requirements.txt
python-telegram-bot[webhooks]==21.5
aiohttp
requests
